Adding controller, service, and route for TipoVehiculo, Puesto, and ServicioOtro modules

+++ src/src/app/vehiculo/route/RutaVehiculo.java
@@ -14,7 +14,7 @@ public class RutaVehiculo {
         return null;
     }
 
-    public void borrarVehiculo(Long codVehiculo) {
+    public void borrarVehiculo(int codVehiculo) {
 
     }
